Extract Habit Log generating into a util
